Resume_QA/requirements.txt
fastapi==0.111.0
pydantic==2.7.1
python-dotenv==1.0.1
uvicorn==0.30.1
langchain==0.1.20
langchain-google-genai==0.0.7